Agregando cambios necesarios para desarrollar HU's 4 y 5

+++ AcessoADatos/Productos/ListarProducto/ListarProductoAD.cs
@@ -26,7 +26,7 @@ namespace AcessoADatos.Productos.ListarProductos
                                                          Producto_ID = elProducto.Producto_ID,
                                                          nombre = elProducto.nombre,
                                                          descripcion = elProducto.descripcion,
-                                                         precio = elProducto.precio,
+                                                         precio = (float)elProducto.precio,
                                                          stock = elProducto.stock,
                                                          imagen = elProducto.imagen,
                                                          Categoria_ID = elProducto.Categoria_ID,
